--- src/main/kotlin/br/unipar/plano/domain/credenciamentos/model/outros/Status.kt
@@ -1,5 +1,0 @@
-package br.unipar.plano.domain.credenciamentos.model.outros
-
-enum class Status {
-        Ativo, Inativo
-}
