+++ src/main/kotlin/br/unipar/plano/domain/credenciamentos/model/status/Status.kt
@@ -1,0 +1,5 @@
+package br.unipar.plano.domain.credenciamentos.model.status
+
+enum class Status {
+        Ativo, Inativo
+}
