+++ src/main/kotlin/br/unipar/plano/domain/credenciamentos/usecases/clinicaHospital/impl/CriaPrestClinHospUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package br.unipar.plano.domain.credenciamentos.usecases.impl.clinicaHospital
+package br.unipar.plano.domain.credenciamentos.usecases.clinicaHospital.impl
 
 import br.unipar.plano.domain.credenciamentos.model.clinicaHospital.PrestadorClinicaHospital
 import br.unipar.plano.domain.credenciamentos.model.clinicaHospital.PrestadorClinicaHospitalRepository
