+++ src/main/kotlin/br/unipar/plano/domain/credenciamentos/usecases/prestadorMedico/impl/CriaPrestMedicoUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package br.unipar.plano.domain.credenciamentos.usecases.impl.prestadorMedico
+package br.unipar.plano.domain.credenciamentos.usecases.prestadorMedico.impl
 
 import br.unipar.plano.domain.credenciamentos.model.prestadorMedico.PrestadorMedico
 import br.unipar.plano.domain.credenciamentos.model.prestadorMedico.PrestadorMedicoRepository
