+++ src/main/kotlin/br/unipar/plano/domain/credenciamentos/usecases/prestadorMedico/impl/DeletaPrestMedicoUseCaseImpl.kt
@@ -1,4 +1,4 @@
-package br.unipar.plano.domain.credenciamentos.usecases.impl.prestadorMedico
+package br.unipar.plano.domain.credenciamentos.usecases.prestadorMedico.impl
 
 import br.unipar.plano.domain.credenciamentos.model.prestadorMedico.IdPrestadorMedico
 import br.unipar.plano.domain.credenciamentos.model.prestadorMedico.PrestadorMedicoRepository
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service
 class DeletaPrestMedicoUseCaseImpl(private val prestadorMedicoRepository: PrestadorMedicoRepository): DeletaPrestMedicoUseCase {
 
     override fun executa(idPrestadorMedico: IdPrestadorMedico){
-        val prestadorMedico = prestadorMedicoRepository.findById(idPrestadorMedico).orElseThrow{ PrestMedicoNotFoundException(idPrestadorMedico)}
+        val prestadorMedico = prestadorMedicoRepository.findById(idPrestadorMedico).orElseThrow{ PrestMedicoNotFoundException(idPrestadorMedico) }
         return prestadorMedicoRepository.delete(prestadorMedico)
     }
